+++ packages/winnow/lib/normalize-query-options/helpers/index.js
@@ -1,4 +1,5 @@
 module.exports = {
   normalizeArray: require('./normalize-array'),
-  detectEsriFieldType: require('./detect-esri-field-type')
+  detectEsriFieldType: require('./detect-esri-field-type'),
+  getCollectionCrs: require('./get-collection-crs')
 }