+++ packages/winnow/lib/normalize-query-options/output-data-spatial-reference.js
@@ -1,4 +1,5 @@
 const normalizeSpatialReference = require('./spatial-reference')
+const { getCollectionCrs } = require('./helpers')
 const logWarning = process.env.NODE_ENV !== 'production' && process.env.KOOP_WARNINGS !== 'suppress'
 
 function normalizeOutputDataSpatialReference (options = {}) {
@@ -9,15 +10,16 @@ function normalizeOutputDataSpatialReference (options = {}) {
     outputCrs,
     outSR,
     inputCrs,
-    sourceSR
+    sourceSR,
+    collection
   } = options
 
   // if no output spatial reference set (outputCrs, projection, srsname, srsName, outSR), assume output will be same as input (inputCrs, sourceSR)
-  const outputSpatialReference = outputCrs || projection || srsname || srsName || outSR || inputCrs || sourceSR || 4326
+  const outputSpatialReference = outputCrs || projection || srsname || srsName || outSR || inputCrs || sourceSR || getCollectionCrs(collection) || 4326
 
   const spatialReference = normalizeSpatialReference(outputSpatialReference)
 
-  if (logWarning) console.log(`WARNING: spatial reference "${outputSpatialReference}" could not be normalized. Defaulting to EPSG:4326.`)
+  if (!spatialReference && logWarning) console.log(`WARNING: spatial reference "${outputSpatialReference}" could not be normalized. Defaulting to EPSG:4326.`)
 
   return spatialReference || { wkid: 4326 }
 }